--- a/src/ggml-cuda/vecdotq.cuh
+++ b/src/ggml-cuda/vecdotq.cuh
@@ -48,8 +48,8 @@ static __device__ __forceinline__ int2 get_int_from_table_16(const int & q4, con
 #define VDR_Q4_0_Q8_1_MMVQ 2
 #define VDR_Q4_0_Q8_1_MMQ  4
 
-// Include GFX906 optimized implementation if available
-#ifdef GGML_HIP_GFX906_OPTIMIZED
+// Include GFX906 optimized implementation
+#if defined(GGML_USE_HIP) || defined(__HIP_PLATFORM_AMD__)
 #include "q4_0-gfx906.cuh"
 #endif
 
@@ -584,7 +584,7 @@ static __device__ __forceinline__ float vec_dot_q6_K_q8_1_impl_mmq(
 static __device__ __forceinline__ float vec_dot_q4_0_q8_1(
     const void * __restrict__ vbq, const block_q8_1 * __restrict__ bq8_1, const int & kbx, const int & iqs) {
 
-#ifdef GGML_HIP_GFX906_OPTIMIZED
+#if defined(GGML_USE_HIP) && defined(__HIP_DEVICE_COMPILE__) && defined(__gfx906__)
     // Use optimized GFX906 implementation with V_DOT8_I32_I4 instruction
     return vec_dot_q4_0_q8_1_gfx906<VDR_Q4_0_Q8_1_MMVQ>(vbq, bq8_1, kbx, iqs);
 #else